Enhance LLVM IR parsing and error reporting
- Updated `parse_llvm_ir` to include a `source_name` parameter for better error messages.
- Improved error handling to include line numbers in exceptions for unsupported LLVM instructions.
- Added tests to verify error reporting and support for `getelementptr` with byte offsets.
- Modified tests to ensure correct behavior with optimized LLVM emission.

+++ src/llvm-to-shasm/tests/fixtures/fib.c
@@ -1,15 +1,15 @@
 #include <stdint.h>
 
 int main(void) {
-    int64_t n = 17;
-    int64_t terms[18];
+    int32_t n = 17;
+    int32_t terms[18];
     terms[0] = 0;
     terms[1] = 1;
 
-    for (int64_t i = 2; i <= n; i++) {
+    for (int32_t i = 2; i <= n; i++) {
         terms[i] = terms[i - 1] + terms[i - 2];
     }
 
-    int64_t out = terms[n];
+    int32_t out = terms[n];
     return (int)out;
 }
